renamed viewModel package
